Fix wonky pyramids
Previously the slope of the pyramid was cast into a double. This adds an
error into the value that is provided as an exact rational number. This
change keeps the value as a rational and only uses the double type where
intermediate values may exceed the integer range.

+++ MCGalaxy/Drawing/DrawOps/AdvPyramidDrawOps.cs
@@ -47,7 +47,7 @@ namespace MCGalaxy.Drawing.Ops {
                 int curHeight = Invert ? yy : height - yy;
                 if (curHeight == 0) continue;              
                 
-                double curRadius = Radius * ((double)curHeight / (double)height);
+                int curRadius = Radius * curHeight / height;
                 if (Math.Abs(xx) > curRadius || Math.Abs(zz) > curRadius) continue;            
                 output(Place(x, y, z, brush));
             }
@@ -78,7 +78,7 @@ namespace MCGalaxy.Drawing.Ops {
                 int curHeight = Invert ? yy : height - yy;
                 if (curHeight == 0) continue;
                 
-                double curRadius = Radius * ((double)curHeight / (double)height);
+                int curRadius = Radius * curHeight / height;
                 int absx = Math.Abs(xx), absz = Math.Abs(zz);
                 if (absx > curRadius || absz > curRadius) continue;
                 if (absx < (curRadius - 1) && absz < (curRadius - 1)) continue;
